Fix typo
I believe that the "scratching" is done by the "curiously beautiful mathematical properties" rather than by the singular "it". So, the properties scratch, not the properties scratches. 

Also, I think the "it does mean it satisfies some curiously..." was too cumbersome, so I removed the second repetition of the word "mean". While we lose the parallel structure, i think the result is more readable.